Fetching categories from DB
