+++ Capstone.Bookstore/Content/MyMainLayout.css
@@ -6,8 +6,8 @@ body {
     margin: 0;
     padding-bottom: 20px;
     font-family: 'Roboto Condensed', sans-serif;
-    font-size: 14px;
-    line-height: 16px;
+    font-size: 16px;
+    line-height: 1.5;
     color: #333;
     background: #fff;
     overflow-x: hidden;
