+++ Capstone.Bookstore/Views/Home/About.cshtml
@@ -1,7 +1,30 @@
 ﻿@{
-    ViewBag.Title = "About";
+    ViewBag.Title = "About Us";
 }
-<h2>@ViewBag.Title.</h2>
-<h3>@ViewBag.Message</h3>
 
-<p>Use this area to provide additional information.</p>
+
+<div class="row">
+    <div class="col-md-12">
+        <div class="page-header">
+            <h2>ABOUT US <small>(Our story)</small></h2>
+        </div>
+    </div>
+</div>
+
+<div class="row">
+    <div class="col-md-12">
+        <p>
+            Book shop is an online bookstore with a mission to financially support local, independent bookstores.
+            We believe that bookstores are essential to a healthy culture. They’re where authors can connect with readers,  where we discover new writers, where children get hooked on the thrill of reading that can last a lifetime. They're also anchors for our downtowns and communities. As more and more people buy their books online, we wanted to create an easy, convenient way for you to get your books and support bookstores at the same time.
+        </p>
+        <p>
+            Today we are the world's largest online independent bookseller, operating with state-of-the-art automation and advanced analytics in a safety-first environment. Our primary customer web and mobile destination is our brand home. We also continue to sell on Amazon, eBay, Abe, and other commerce sites where book lovers gather. We hand-grade every book in our inventory and offer books in a variety of formats and conditions for our customers. We pride ourselves on our wide assortment of quality, accurately graded books provided to our customers at everyday low prices. Our award-winning customer service team and our 5-star rating on Trustpilot, with over 640,000 reviews, provide our customers' assurance to shop with confidence while finding books they love and look forward to reading. Our Reading Rewards loyalty program continually delights loyal customers with free books awarded for their points earned.
+        </p>
+        <p>
+            Our partnerships extend into communities across the country. Our profit-sharing relationship with libraries large and small enables us to support these valuable institutions by finding homes for ex-library books and working with Friends of Library groups across the country to provide financial support for their continued success. We give back to our communities through our Books Cares pillar, designed to connect with communities in the US and also­ globally donating books to those in need—including schools, non-profit organizations, and those rebuilding after natural disasters.
+        </p>        
+        <p>
+            We believe in books and the power of the written word to impact lives and societies across the globe, and we cherish our role in saving books to live another day to educate, entertain, inspire, and transform lives.
+        </p>
+    </div>
+</div>
